requirements.txt
vines-worker-sdk==0.0.19
langchain
load
python-dotenv
paddleocr
paddlepaddle
selenium